ISPN-8259 Test field @Analyzer for Protobuf
