+++ client/hotrod-client/src/test/java/org/infinispan/client/hotrod/query/RemoteQueryStringTest.java
@@ -16,7 +16,9 @@ import org.infinispan.client.hotrod.RemoteCacheManager;
 import org.infinispan.client.hotrod.Search;
 import org.infinispan.client.hotrod.exceptions.HotRodClientException;
 import org.infinispan.client.hotrod.marshall.ProtoStreamMarshaller;
+import org.infinispan.client.hotrod.query.testdomain.protobuf.AnalyzerTestEntity;
 import org.infinispan.client.hotrod.query.testdomain.protobuf.ModelFactoryPB;
+import org.infinispan.client.hotrod.query.testdomain.protobuf.marshallers.AnalyzerTestEntityMarshaller;
 import org.infinispan.client.hotrod.query.testdomain.protobuf.marshallers.MarshallerRegistration;
 import org.infinispan.client.hotrod.query.testdomain.protobuf.marshallers.NotIndexedMarshaller;
 import org.infinispan.client.hotrod.test.HotRodClientTestingUtil;
@@ -33,6 +35,7 @@ import org.infinispan.query.dsl.embedded.testdomain.User;
 import org.infinispan.query.remote.client.ProtobufMetadataManagerConstants;
 import org.infinispan.server.hotrod.HotRodServer;
 import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 /**
@@ -49,10 +52,31 @@ public class RemoteQueryStringTest extends QueryStringTest {
          "message NotIndexed {\n" +
          "\toptional string notIndexedField = 1;\n" +
          "}\n";
+
+   private static final String CUSTOM_ANALYZER_PROTO_SCHEMA = "package sample_bank_account;\n" +
+           "/* @Indexed \n" +
+           "   @Analyzer(definition = \"standard\") */" +
+           "message AnalyzerTestEntity {\n" +
+           "\t/* @Field(store = Store.YES, analyze = Analyze.YES, analyzer = @Analyzer(definition = \"stemmer\")) */\n" +
+           "\toptional string f1 = 1;\n" +
+           "\t/* @Field(store = Store.YES, analyze = Analyze.NO, indexNullAs = \"-1\") */\n" +
+           "\toptional int32 f2 = 2;\n" +
+           "}\n";
+
    protected HotRodServer hotRodServer;
    protected RemoteCacheManager remoteCacheManager;
    protected RemoteCache<Object, Object> remoteCache;
    protected Cache<Object, Object> cache;
+
+   @BeforeClass
+   @Override
+   protected void populateCache() throws Exception {
+      super.populateCache();
+
+      getCacheForWrite().put("analyzed1", new AnalyzerTestEntity("tested 123", 3));
+      getCacheForWrite().put("analyzed2", new AnalyzerTestEntity("testing 1234", 3));
+      getCacheForWrite().put("analyzed3", new AnalyzerTestEntity("xyz", null));
+   }
 
    @Override
    protected QueryFactory getQueryFactory() {
@@ -95,13 +119,16 @@ public class RemoteQueryStringTest extends QueryStringTest {
       RemoteCache<String, String> metadataCache = remoteCacheManager.getCache(ProtobufMetadataManagerConstants.PROTOBUF_METADATA_CACHE_NAME);
       metadataCache.put("sample_bank_account/bank.proto", Util.read(Util.getResourceAsStream("/sample_bank_account/bank.proto", getClass().getClassLoader())));
       metadataCache.put("not_indexed.proto", NOT_INDEXED_PROTO_SCHEMA);
+      metadataCache.put("custom_analyzer.proto", CUSTOM_ANALYZER_PROTO_SCHEMA);
       assertFalse(metadataCache.containsKey(ProtobufMetadataManagerConstants.ERRORS_KEY_SUFFIX));
 
       //initialize client-side serialization context
       SerializationContext serCtx = ProtoStreamMarshaller.getSerializationContext(remoteCacheManager);
       MarshallerRegistration.registerMarshallers(serCtx);
       serCtx.registerProtoFiles(FileDescriptorSource.fromString("not_indexed.proto", NOT_INDEXED_PROTO_SCHEMA));
+      serCtx.registerProtoFiles(FileDescriptorSource.fromString("custom_analyzer.proto", CUSTOM_ANALYZER_PROTO_SCHEMA));
       serCtx.registerMarshaller(new NotIndexedMarshaller());
+      serCtx.registerMarshaller(new AnalyzerTestEntityMarshaller());
    }
 
    protected ConfigurationBuilder getConfigurationBuilder() {
@@ -170,5 +197,14 @@ public class RemoteQueryStringTest extends QueryStringTest {
 
       List<User> list = q.list();
       assertEquals(3, list.size());
+   }
+
+   public void testCustomFieldAnalyzer() throws Exception {
+      QueryFactory qf = getQueryFactory();
+
+      Query q = qf.create("from sample_bank_account.AnalyzerTestEntity where f1:'test'");
+
+      List<AnalyzerTestEntity> list = q.list();
+      assertEquals(2, list.size());
    }
 }
